Broken build! Added pyproject structure, updated folder structure, updated shared_info concept, added test folder
