+++ src/backend/networking-module/main.py
@@ -1,0 +1,8 @@
+from networking_lib.shared_info import Info
+
+def main():
+    pass
+
+
+if __name__ == '__main__':
+    main()
