+++ src/backend/networking-module/pyproject.toml
@@ -1,0 +1,15 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "networking_lib"
+version = "0.1.0"
+description = "A modular networking library for probing"
+dependencies = [
+    "socket",
+    "re",
+]
+
+[tool.setuptools.packages.find]
+where = ["src"]
