+++ src/backend/networking-module/tests/dns_resolver_test.py
@@ -1,0 +1,5 @@
+import "../shared_info.py"
+
+s = shared_info()
+
+print(s)
